# Optimize GetVisualTreeElementsWindowsInternal