There were multiple enumerations occurring in the Windows implementation of GetVisualTreeElements.

Also, using a HashSet, in particular, for the Contains() check within the other enumeration should especially help speed things up.