@symbiogenesis 2024-01-19
Added a benchmark.

Before:
```
| Method                | Mean     | Error   | StdDev  | Gen0      | Gen1      | Gen2     | Allocated |
|---------------------- |---------:|--------:|--------:|----------:|----------:|---------:|----------:|
| GetVisualTreeElements | 134.6 ms | 2.64 ms | 2.60 ms | 6750.0000 | 2000.0000 | 250.0000 |  58.46 MB |
```


After:
```
| Method                | Mean     | Error   | StdDev  | Gen0      | Gen1      | Gen2     | Allocated |
|---------------------- |---------:|--------:|--------:|----------:|----------:|---------:|----------:|
| GetVisualTreeElements | 125.9 ms | 2.11 ms | 1.76 ms | 6600.0000 | 1800.0000 | 200.0000 |  58.46 MB |
```