@symbiogenesis 2024-01-19
Latest version is about the same performance.

| Method                | Mean     | Error   | StdDev  | Gen0      | Gen1      | Gen2     | Allocated |
|---------------------- |---------:|--------:|--------:|----------:|----------:|---------:|----------:|
| GetVisualTreeElements | 126.5 ms | 1.88 ms | 1.67 ms | 6750.0000 | 2000.0000 | 250.0000 |  58.46 MB |