@symbiogenesis 2024-01-19
Benchmarks for latest version:

| Method                | Mean     | Error   | StdDev  | Gen0      | Gen1      | Gen2     | Allocated |
|---------------------- |---------:|--------:|--------:|----------:|----------:|---------:|----------:|
| GetVisualTreeElements | 123.2 ms | 2.32 ms | 2.27 ms | 6800.0000 | 1800.0000 | 400.0000 |  58.46 MB |